Refactor pwrsupply subpackage
