+++ siriuspy/siriuspy/devices/dcct.py
@@ -2,7 +2,7 @@
 
 import time as _time
 
-from ..pwrsupply.status import PSCStatus as _PSCStatus
+from ..pwrsupply.psctrl.pscstatus import PSCStatus as _PSCStatus
 
 from .device import Device as _Device
