+++ siriuspy/siriuspy/pwrsupply/bsmp/__init__.py
@@ -1,0 +1,1 @@
+"""Power Supply BSMP."""
