+++ siriuspy/siriuspy/search/ll_time_search.py
@@ -4,12 +4,13 @@ import re as _re
 from copy import deepcopy as _dcopy
 from threading import Lock as _Lock
 
-from siriuspy import clientweb as _web
-from siriuspy.namesys import SiriusPVName as _PVName, Filter as _Filter
+from .. import clientweb as _web
+from ..namesys import SiriusPVName as _PVName, Filter as _Filter
+
 from .bpms_search import BPMSearch as _BPMSearch
 from .ps_search import PSSearch as _PSSearch
 
-_timeout = 1.0
+_TIMEOUT = 1.0
 
 
 class LLTimeSearch:
@@ -326,7 +327,7 @@ class LLTimeSearch:
                 return
             if not _web.server_online():
                 raise Exception('Could not connect with Consts Server!!')
-            text = _web.timing_devices_mapping(timeout=_timeout)
+            text = _web.timing_devices_mapping(timeout=_TIMEOUT)
             cls._parse_text_and_build_connection_mappings(text)
             cls._update_related_maps()
